+++ src/components/Prayers/Prayers.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import Prayers from './Prayers';
 
-describe('<App />', () => {
+describe('<Prayers />', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div');
-    ReactDOM.render(<App />, div);
+    ReactDOM.render(<Prayers />, div);
     ReactDOM.unmountComponentAtNode(div);
   });
 })